# [CANN]get_rows and dup optimization