… ggml/src/ggml-cann/kernels/CMakeLists.txt
… ggml/src/ggml-cann/kernels/ascendc_kernels.h
… ggml/src/ggml-cann/kernels/dup.cpp
… ggml/src/ggml-cann/kernels/get_row_f16.cpp
… ggml/src/ggml-cann/kernels/get_row_f32.cpp
… ggml/src/ggml-cann/kernels/get_row_q4_0.cpp
… ggml/src/ggml-cann/kernels/get_row_q8_0.cpp
… ggml/src/ggml-cann/kernels/quantize_f16_q8_0.cpp
… ggml/src/ggml-cann/kernels/quantize_f32_q8_0.cpp
… ggml/src/ggml-cann/kernels/quantize_float_to_q4_0.cpp