### Why is this PR needed?
Optimize the `get_rows` and `dup` operator in the CANN backend by replacing the previous AscendC implementation with the aclnn acceleration library.

Backend 2/2: CPU
  Skipping
2/2 backends passed
OK


### Inference Testing
**Model**
Qwen-0.5B
**Script**
```
./bin/llama-cli -m /home/lcg/gguf_model/Qwen2-0___5B-Instruct.gguf -p "Building a website can be done in 10 steps:" -ngl 32
```
**Before**
```
llama_perf_sampler_print:    sampling time =     703.64 ms /   410 runs   (    1.72 ms per token,   582.68 tokens per second)
llama_perf_context_print:        load time =    4759.33 ms
llama_perf_context_print: prompt eval time =      29.71 ms /    31 tokens (    0.96 ms per token,  1043.28 tokens per second)
llama_perf_context_print:        eval time =    8208.92 ms /   378 runs   (   21.72 ms per token,    46.05 tokens per second)
llama_perf_context_print:       total time =   10230.50 ms /   409 tokens
```
**After**
```
llama_perf_sampler_print:    sampling time =     592.66 ms /   352 runs   (    1.68 ms per token,   593.93 tokens per second)
llama_perf_context_print:        load time =    4545.10 ms
llama_perf_context_print: prompt eval time =      25.05 ms /    31 tokens (    0.81 ms per token,  1237.52 tokens per second)
llama_perf_context_print:        eval time =    6513.81 ms /   320 runs   (   20.36 ms per token,    49.13 tokens per second)
llama_perf_context_print:       total time =    8978.30 ms /   351 tokens
```